Split the pipes unit tests into their own related submodule
