+++ scanpipe/tests/pipes/test_resolve.py
@@ -1,0 +1,44 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# http://nexb.com and https://github.com/nexB/scancode.io
+# The ScanCode.io software is licensed under the Apache License version 2.0.
+# Data generated with ScanCode.io is provided as-is without warranties.
+# ScanCode is a trademark of nexB Inc.
+#
+# You may not use this software except in compliance with the License.
+# You may obtain a copy of the License at: http://apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+#
+# Data Generated with ScanCode.io is provided on an "AS IS" BASIS, WITHOUT WARRANTIES
+# OR CONDITIONS OF ANY KIND, either express or implied. No content created from
+# ScanCode.io should be considered or used as legal advice. Consult an Attorney
+# for any legal advice.
+#
+# ScanCode.io is a free software code scanning tool from nexB Inc. and others.
+# Visit https://github.com/nexB/scancode.io for support and download.
+
+from django.test import TestCase
+
+from scanpipe.pipes import resolve
+
+
+class ScanPipeResolvePipesTest(TestCase):
+    def test_scanpipe_pipes_resolve_set_license_expression(self):
+        declared_license = {"license": "MIT"}
+        data = resolve.set_license_expression({"declared_license": declared_license})
+        self.assertEqual("mit", data.get("license_expression"))
+
+        declared_license = {
+            "classifiers": [
+                "License :: OSI Approved :: Python Software Foundation License"
+            ]
+        }
+        data = resolve.set_license_expression({"declared_license": declared_license})
+        self.assertEqual("python", data.get("license_expression"))
+
+        declared_license = "GPL 2.0"
+        data = resolve.set_license_expression({"declared_license": declared_license})
+        self.assertEqual("gpl-2.0", data.get("license_expression"))
